+++ Example App/lib/src/Core/Landing/Resource Generation/GenerationLandingView.dart
@@ -51,17 +51,6 @@ class GenerationLandingPage extends StatelessWidget {
       throw ('Unexpected platform brightness issue. Please check the implementation.');
     }
 
-    Image landscapeBackgroundImage() {
-      if (palette.brightness() == Brightness.light) {
-        return Image.network(
-            'https://images.unsplash.com/photo-1526934709557-35f3777499c4?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2069&q=80');
-      } else if (palette.brightness() == Brightness.dark) {
-        return Image.network(
-            'https://images.unsplash.com/photo-1471047283799-ebd97acc0bc3?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2577&q=80');
-      }
-      throw ('Unexpected platform brightness issue. Please check the implementation.');
-    }
-
     return Container();
   }
 }